Devide tests into tests for old registration frontend (without gateway) = BrowserRegistrationSpec and tests for new setup process (with gateway).
